+++ exts/omni.isaac.lab_quadruped_tasks/omni/isaac/lab_quadruped_tasks/mdp/actions/__init__.py
@@ -1,0 +1,8 @@
+"""
+License: MIT License
+Copyright (c) 2025, Felipe Mohr Santos
+"""
+
+
+from .actions_cfg import *
+from .quadruped_actions import *